Refacto classe PaymentSrtviceImpl
