+++ src/main/java/com/payline/payment/docapost/service/impl/PaymentFormConfigurationServiceImpl.java
@@ -16,7 +16,6 @@ import org.apache.logging.log4j.Logger;
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
 import java.io.ByteArrayOutputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Locale;
@@ -30,8 +29,6 @@ public class PaymentFormConfigurationServiceImpl implements PaymentFormConfigura
     private static final int LOGO_WIDTH = 200;//141
     private static final Logger LOGGER = LogManager.getLogger(PaymentFormConfigurationServiceImpl.class);
     private I18nService i18n = I18nService.getInstance();
-
-
 
     @Override
     public PaymentFormConfigurationResponse getPaymentFormConfiguration(PaymentFormConfigurationRequest paymentFormConfigurationRequest) {
